Implement functions for birthday calculations
Implement the function get_birthday() that returns a dictionary
containing a list of three integers representing the year, month
and the day of birth, respectively. The list is saved under the
key which is the name of employee.

Implement the function is_leap_year() that returns True if the
year is leap or False if it is not.